+++ reactionmenu/errors.py
@@ -24,18 +24,25 @@ DEALINGS IN THE SOFTWARE.
 
 import inspect
 
-class ViewMenuException(Exception):
+class MenuException(Exception):
+	"""Base exception for all menu's
+	
+		.. added:: v3.0.0
+	"""
+	pass
+
+class ViewMenuException(MenuException):
 	"""Base :class:`ViewMenu` exception
 	
 		.. added:: v3.0.0
 	"""
 	pass
 
-class ReactionMenuException(Exception):
+class ReactionMenuException(MenuException):
 	"""Base :class:`ReactionMenu` exception"""
 	pass
 
-class IncorrectType(ReactionMenuException):
+class IncorrectType(MenuException):
 	"""Raised when the expected type was not given
 	
 		.. added:: v1.0.9
@@ -43,7 +50,7 @@ class IncorrectType(ReactionMenuException):
 	def __init__(self, message: str):
 		super().__init__(message)
 
-class NoButtons(ReactionMenuException):
+class NoButtons(MenuException):
 	"""Raised when the menu was started but no buttons were registered or the action initiated requires buttons to be registered
 
 		.. added:: v1.0.9
@@ -51,63 +58,50 @@ class NoButtons(ReactionMenuException):
 	def __init__(self, message: str='You cannot start the menu when no buttons are registered'):
 		super().__init__(message)
 
-class InvalidPage(ReactionMenuException):
+class InvalidPage(MenuException):
 	"""Raised when the page selected to remove does not exist"""
 	pass
 
-class SingleUseOnly(ReactionMenuException):
+class SingleUseOnly(MenuException):
 	"""When a method is called more than once"""
 	pass
 	
-class DescriptionOversized(ReactionMenuException):
+class DescriptionOversized(MenuException):
 	"""The content of the menu has a character size > 4096"""
 	pass
 
-class MenuSettingsMismatch(ReactionMenuException):
+class MenuSettingsMismatch(MenuException):
 	"""Used settings for a static menu but is using dynamic methods and vice-versa"""
 	pass
 
-class DuplicateButton(ReactionMenuException):
+class DuplicateButton(MenuException):
 	"""The emoji selected as a button has already been registered"""
 	pass
 
-class ImproperStyleFormat(ReactionMenuException):
+class ImproperStyleFormat(MenuException):
 	"""The custom style selected by the user did not meet formatting requirements"""
 	def __init__(self):
 		super().__init__('There needs to be at least but no more than 1 "$" character and at least but no more than 1 "&" character. "$" represents the current page, "&" represents the total amount of pages. Example: "On $ out of &"')
 
-class ImproperButtonOrderChange(ReactionMenuException):
-	"""Buttons were leftout/added when changing the order"""
-	def __init__(self, missing, extra):
-		super().__init__(inspect.cleandoc(f"""Missing or extra button(s) when changing the order. The button order list must match the buttons that were already added. You cannot add or leave out any buttons
-
-			Missing = {missing if missing else "<none missing>"} | Extras = {extra if extra else "<no extras>"}
-
-			NOTE: If the menu only has one page, you still need to include the "next_button" and "back_button" in the order, because those are there by default. Add them with.. 
-			1 - Attributes ReactionMenu.default_back_button and ReactionMenu.default_next_button. 
-			2 - Method ReactionMenu.help_appear_order can be used to print all emojis from the Buttons you've added to the python console for you to copy and paste in the desired order.
-			3 - Method ReactionMenu.get_button_by_name to get the button object
-		"""))
-
-class TooManyButtons(ReactionMenuException):
+class TooManyButtons(MenuException):
 	"""The amount of buttons registered is > 20 (over discords reaction limit)"""
 	def __init__(self, message: str='Discord currently has a limit of 20 reactions or less per message. Remove 1 or more buttons'):
 		super().__init__(message)
 
-class MissingSetting(ReactionMenuException):
+class MissingSetting(MenuException):
 	"""Raised when an action requires specific input from the user"""
 	pass
 
-class NoPages(ReactionMenuException):
+class NoPages(MenuException):
 	"""Tried to start the menu when they haven't added any pages"""
 	def __init__(self, message: str="You cannot start a reaction menu when there aren't any pages"):
 		super().__init__(message)
 
-class MenuAlreadyRunning(ReactionMenuException):
+class MenuAlreadyRunning(MenuException):
 	"""Called a method that is not allowed to be called after the menu has started"""
 	def __init__(self, message: str):
 		super().__init__(message)
 
-class ButtonNotFound(ReactionMenuException):
+class ButtonNotFound(MenuException):
 	"""Raised when the :meth:ReactionMenu.remove_button did not find any matching buttons"""
 	pass